refactor some data structures

+++ Assets/PurrNet/Runtime/Components/NetworkIdentity/NetworkIdentity.Modules.cs
@@ -43,7 +43,7 @@ namespace PurrNet
             _externalModulesView.Add(module);
         }
 
-        public bool TryGetModule(byte moduleId, out NetworkModule module)
+        public bool TryGetModule(int moduleId, out NetworkModule module)
         {
             if (moduleId >= _modules.Count)
             {
